refactor(api): move OpenAPI endpoints to FrontendModule
Move /openapi.json, /docs, and /swagger endpoints from configure.swift
to FrontendModule for better code organization and separation of concerns.

Changes:
- Move OpenAPI specification endpoint to FrontendRouter
- Move Swagger UI serving endpoint to FrontendRouter
- Move /swagger redirect to FrontendRouter
- Exclude OpenAPI endpoints from appearing in the spec itself
- Keep /health endpoint in configure.swift as infrastructure concern

Addresses review feedback from PR #22.
